% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorize.R
\name{categorize}
\alias{categorize}
\title{This function uses pre-given data sets to create categorical data sets for given thresholds.}
\usage{
categorize(f.loc, threshold, dataList = "Data_List.dat")
}
\arguments{
\item{f.loc}{File location. Generated data sets will be saved at the user-defined location.}

\item{threshold}{The threshold values.}

\item{dataList}{List of the names of data sets generated earlier either with the package functions or any other software.}
}
\description{
Previously simulated data sets are utilized to create categorical data sets by the given thresholds.
}
\examples{
tres<-c(-Inf, -1.645, -.643, .643, 1.645, Inf) # five categories
categorize(f.loc=tempdir(), threshold = tres)
}
\author{
Fatih Orçan
}
